% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_FTE_value.R
\name{sim_fte_value}
\alias{sim_fte_value}
\title{Title}
\usage{
sim_fte_value(n_sim, min_value, max_value, min_share)
}
\arguments{
\item{n_sim}{Nr of values to simulate.}

\item{min_value}{Minimum value of FTE.}

\item{max_value}{Maximum value of FTE.}

\item{min_share}{Share of minimal values.}
}
\value{
A numeric vector.
}
\description{
Title
}
\examples{
sim_fte_value(12, 5000, 24000, 0.5)
}
