% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_jobs_nr.R
\name{sim_jobs_nr}
\alias{sim_jobs_nr}
\title{Title}
\usage{
sim_jobs_nr(n_sim, jobs_mean_month)
}
\arguments{
\item{n_sim}{Number of months to simulate.}

\item{jobs_mean_month}{Mean jobs per month.}
}
\value{
A numeric vector of n_sim size.
}
\description{
Title
}
\examples{
sim_jobs_nr(60, 20)
}
