% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_monthly_eff.R
\name{sim_monthly_eff}
\alias{sim_monthly_eff}
\title{Title}
\usage{
sim_monthly_eff(n_sim, eff_min, eff_max)
}
\arguments{
\item{n_sim}{Number of months to simulate.}

\item{eff_min}{Minimum monthly efficiency.}

\item{eff_max}{Maximum monthly efficiency.}
}
\value{
A numeric vector with values between 0 and 1.
}
\description{
Title
}
\examples{
sim_monthly_eff(12, 0.2, 0.5)
}
