% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_nr_of_FTE_per_job.R
\name{sim_nr_of_FTE_per_job}
\alias{sim_nr_of_FTE_per_job}
\title{Title}
\usage{
sim_nr_of_FTE_per_job(n_sim, min_fte, max_fte, min_share)
}
\arguments{
\item{n_sim}{Nr of months to simulate.}

\item{min_fte}{Minimum FTE per process.}

\item{max_fte}{Maximum FTE per process.}

\item{min_share}{Share of low FTE processes.}
}
\value{
A numeric vector.
}
\description{
Title
}
\examples{
sim_nr_of_FTE_per_job(12, 1,10,0.5)
}
