% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_results.R
\name{sim_results}
\alias{sim_results}
\title{Title}
\usage{
sim_results(
  n_sim_jobs,
  jobs_mean_month,
  eff_min,
  eff_max,
  min_fte,
  max_fte,
  min_share_fte,
  min_value,
  max_value,
  min_share_value
)
}
\arguments{
\item{n_sim_jobs}{Number of months to simulate.}

\item{jobs_mean_month}{Number of jobs proccesed monthly.}

\item{eff_min}{Minimum efficiency per job (FTE confirmed / FTE goal).}

\item{eff_max}{Maximum efficiency per job (FTE confirmed / FTE goal).}

\item{min_fte}{Minimum FTEs per job.}

\item{max_fte}{Maximum FTEs per job.}

\item{min_share_fte}{Share of jobs with minimal or near number of FTEs.}

\item{min_value}{Minimum value of one FTE.}

\item{max_value}{Maximum value of one FTE.}

\item{min_share_value}{Share of FTEs valued near minimum.}
}
\value{
A tibble with simulated data for each month, job and other factors.
}
\description{
Title
}
\examples{
sim_results(n_sim_jobs = 12, jobs_mean_month = 20,  eff_min = 0.2, eff_max = 0.4, min_fte = 1, max_fte = 5, min_share_fte = 0.5, min_value = 5000, max_value = 15000, min_share_value = 0.5)
}
